normal dist working

+++ src_ori/registry_cia.py
@@ -125,7 +125,7 @@ def _load_cia_npz(index: int, path: str, target_wavelengths: np.ndarray) -> CiaR
         idx=index,
         temperatures=jnp.asarray(temperatures),
         wavelengths=jnp.asarray(target_wavelengths),
-        cross_sections=jnp.asarray(xs_interp),
+        cross_sections=jnp.asarray(xs_interp, dtype=jnp.float32),
     )
 
 # Pad the tables to a rectangle (in dimension) - usually only in T as wavelength grids are the same lengths
@@ -203,7 +203,7 @@ def _build_hminus_cia_entry(index: int, target_wavelengths: np.ndarray, spec) ->
         idx=index,
         temperatures=jnp.asarray(T),
         wavelengths=jnp.asarray(lam),
-        cross_sections=jnp.asarray(log10_sigma),
+        cross_sections=jnp.asarray(log10_sigma, dtype=jnp.float32),
     )
